Update: Ensure docker build fails without `AUTH_TOKEN` build arg
Since the `GITHUB_TOKEN` being used as the environment variable in
the docker image was needed (and not optional) - allowing docker
builds to succeed, only to result in a crash when the container was
run would be a waste of time.

With this commit, the `AUTH_TOKEN` argument is mandatory for the
`docker build` command to succeed. This greatly cuts down the
response time to detect an error with the build command